ビルドツールにボタンを追加できるボタンプラグインの仕組みを実装

+++ phest/plugins/phest/s3deploy.php
@@ -1,9 +1,9 @@
 <?php
     namespace ChatWork\Phest;
-    use \ChatWork\Utility\File;
-    use \Aws\S3\S3Client;
-    use \Aws\S3\Enum\CannedAcl;
-    use \Guzzle\Http\EntityBody;
+    use ChatWork\Utility\File;
+    use Aws\S3\S3Client;
+    use Aws\S3\Enum\CannedAcl;
+    use Guzzle\Http\EntityBody;
 
 /**
  * Amazon S3へProductionビルドをアップロード
@@ -42,47 +42,26 @@
  * @param  string  $params.secret AWSのシークレットキー
  * @param  Phest $phest Phestオブジェクト
  */
-function plugin_finish_s3deploy(array $params, Phest $phest){
-    if ($phest->getBuildType() != 'production'){
-        return false;
-    }
+function plugin_button_s3deploy(array $params, Phest $phest){
+    $phest->registerSection('s3deploy','デプロイ完了');
+    $phest->registerSection('s3deployerror','デプロイエラー',array('type' => 'danger'));
 
     require(DIR_PHEST.'/lib/vendor/aws/aws.phar');
 
-    $bucket = $params['bucket'];
-    $region_key = strtoupper($params['region']);
-    $region = constant('\Aws\Common\Enum\Region::'.$region_key);
-    $access_key = $params['key'];
-    $secret_key = $params['secret'];
+    if ($phest->hasError()){
+        $phest->add('s3deployerror','エラーが発生しているためデプロイしません');
+        return false;
+    }
+
+    $dryrun = true;
+    if (isset($params['dryrun'])){
+        $dryrun = $params['dryrun'];
+    }
+
     $prefix = '';
     if (isset($params['prefix'])){
         $prefix = $params['prefix'];
     }
-
-    $enable = $params['enable'];
-
-    if (!$enable){
-        $phest->add('build','[deploy] enableがfalseになっているためデプロイしません');
-        return false;
-    }
-
-    if ($phest->hasError()){
-        $phest->add('build','[deploy] エラーが発生しているためデプロイしません');
-        return false;
-    }
-
-    if (!$bucket or !$region or !$access_key or !$secret_key){
-        die('deployの指定が間違っています');
-    }
-
-    $phest->registerSection('deploy','デプロイ完了 Region: '.$region.' Bucket: '.$bucket.' Prefix: '.$prefix);
-    $phest->registerSection('deployerror','デプロイエラー',array('type' => 'danger'));
-
-    $s3 = S3Client::factory(array(
-        'key' => $access_key,
-        'secret' => $secret_key,
-        'region' => $region
-        ));
 
     //$buckets = $s3->listBuckets();
     //print_a($buckets);
@@ -92,22 +71,69 @@ function plugin_finish_s3deploy(array $params, Phest $phest){
     //  print_a($object);
     // }
 
-    $upload_dir = $phest->getOutputPath();
+    $upload_dir = $phest->getOutputPath($phest->getLang(),'production');
+
+    if (!is_dir($upload_dir)){
+        $phest->add('s3deployerror','productionでビルドされたファイルが存在しません');
+        return false;
+    }
+
+
+    if (!$dryrun){
+        $is_error = false;
+        $required = array('bucket','region','key','secret');
+        foreach ($required as $rval){
+            if (empty($params[$rval])){
+                $phest->add('s3deployerror','オプションが指定されていません: '.$rval);
+                $is_error = true;
+            }
+        }
+
+        if ($is_error){
+            return false;
+        }
+
+        $bucket = $params['bucket'];
+        $region_key = strtoupper($params['region']);
+        $region = constant('\Aws\Common\Enum\Region::'.$region_key);
+        $access_key = $params['key'];
+        $secret_key = $params['secret'];
+
+        $phest->add('s3deploy','リージョン: '.$region.' バケット: '.$bucket);
+        try {
+            $s3 = S3Client::factory(array(
+                'key' => $access_key,
+                'secret' => $secret_key,
+                'region' => $region
+                ));
+        } catch (\Exception $e){
+            $phest->add('s3deployerror',$e->getMessage());
+            return false;
+        }
+    }
+
     $file_list = File::getFileList($upload_dir);
     foreach ($file_list as $filepath){
         $key = strtr(ltrim(substr($filepath,strlen($upload_dir)),'\\/'),'\\','/');
 
         //upload
         try {
-            $result = $s3->putObject(array(
-                'Bucket' => $bucket,
-                'Key' => $prefix.$key,
-                'Body' => EntityBody::factory(fopen($filepath,'r+')),
-                'ACL' => CannedAcl::PUBLIC_READ,
-                ));
-            $phest->add('deploy',$prefix.$key);
+            if (!$dryrun){
+                $result = $s3->putObject(array(
+                    'Bucket' => $bucket,
+                    'Key' => $prefix.$key,
+                    'Body' => EntityBody::factory(fopen($filepath,'r+')),
+                    'ACL' => CannedAcl::PUBLIC_READ,
+                    ));
+                $phest->add('s3deploy',$prefix.$key);
+            }else{
+                $phest->add('s3deploy',$prefix.$key.' <code>[dryrun]</code>');
+            }
         } catch (S3Exception $e){
-            $phest->add('deployerror',$e->getMessage());
+            $phest->add('s3deployerror',$e->getMessage());
+        } catch (\Exception $e){
+            $phest->add('s3deployerror',$e->getMessage());
+            return false;
         }
     }
 }
